refac serviços compilados em classes + correções
